[코드 정리] 불필요한 코드 제거 및 확장 메소드 추가

+++ Assets/Script/GameElement/ItemEvent.cs
@@ -2,8 +2,11 @@
 using System.Collections.Generic;
 using UnityEngine;
 using System;
+using GGame;
 
-// 이 스크립트는 사용하지 않음 -> ItemScript로 대체
+/// <summary>
+/// 해당 스크립트는 사용하지 않음 -> ItemScript로 대체 
+/// </summary>
 
 public class ItemEvent : MonoBehaviour
 {
@@ -12,11 +15,10 @@ public class ItemEvent : MonoBehaviour
     //아이템 충돌 처리
     void OnTriggerEnter2D(Collider2D collision)
     {
-        if (collision.transform.tag.ToUpper().Equals(Enum.GetName(typeof(ECharacter), (int)ECharacter.PAC)))
+        if (collision.transform.tag.ToUpper().Equals(Enum.GetName(typeof(ECharacter), ECharacter.PAC)))
         {
             ItemBuff(); 
             gameObject.SetActive(false);
-
         }
     }
